Create README.md
This project consists of a part of the functionalities of a virtual store, more specifically, list of products, categories and shopping cart. The store's product categories are shown on the home page, the user clicks on one of the options and opens the page with the list of products in that category, in each of which there is the option to like the product and add it to the shopping cart. In the navigation bar there is a search field where the list of products will be filtered according to the keyword typed, beside a cart icon that, when clicked, will take the user to the shopping cart page with the products already added. In the shopping cart, a list of all products added is shown, with the possibility of changing the quantity of each item, at the bottom of the page, the subtotal is shown with the sum of the prices of all products and, just below, a button to complete the purchase, which in this example resets the shopping cart.

The main purpose of this project is the use of Redux (react-redux) to manage global states. Among other libraries and technologies used are react-multi-carousel and the styles part with Sass.